rename common to shared in agents
