+++ agent_app_sample_code/cookbook/config/shared/agent_storage_location.py
